habitat_requirements.txt
absl-py==2.0.0
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1683424013410/work
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
braceexpand==0.1.7
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1693583441880/work
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.5
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.1.4
configparser==6.0.0
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.1
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1669279893622/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1694522931813/work
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
gym==0.26.2
gym-notices==0.0.8
-e git+https://github.com/yashkant/habitat-lab-fork.git@df9753bf8353ac11438fb80114528fa9f0c1945e#egg=habitat&subdirectory=../../habitat-lab
habitat-sim==0.2.1
huggingface-hub==0.17.2
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.6.1
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1694632425602/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==6.1.0
ipdb==0.13.13
ipykernel==6.25.2
ipython==8.12.2
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json-tricks==3.17.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1695379923772/work
libtmux==0.23.2
lit==16.0.6
llvmlite @ file:///croot/llvmlite_1683555126642/work
lmdb==1.4.1
magnum @ file:///opt/conda/envs/py36/conda-bld/habitat-sim_1627788414467/work/build/deps/magnum-bindings/src/python
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-base_1594091694890/work
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook_shim==0.2.3
numba @ file:///home/conda/feedstock_root/build_artifacts/numba_1687804751716/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808301083/work
numpy-quaternion @ file:///home/conda/feedstock_root/build_artifacts/quaternion_1676352986710/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
objectio==0.2.29
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
openai==0.27.9
opencv-python==4.8.0.76
ortools==8.0.8283
overrides==7.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pkgutil_resolve_name==1.3.10
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
prometheus-client==0.17.1
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.0.4
pycparser==2.21
pygame==2.5.2
Pygments==2.16.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1690737849915/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==5.4.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scikit-fmm==2023.4.2
scikit-learn==1.3.1
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1683900206454/work/base/dist/scipy-1.10.1-cp38-cp38-linux_x86_64.whl#sha256=13ab5c17dedeb97a65ba90de1c51475f2d3ae0fe50bc43e9b018cc2943541de4
seaborn==0.10.1
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.31.0
shapely==2.0.1
shortuuid==1.0.11
simplejson==3.19.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1611376390914/work
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
subprocess32==3.5.4
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1695373547159/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1691671248568/work
traitlets==5.10.0
transformers==4.33.0
trimesh==3.23.5
triton==2.0.0
typer==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wandb==0.10.11
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdataset==0.1.40
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.7
yacs==0.1.8
yarl==1.9.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
